Implemented downsampling matrix getter.
Test passed.

+++ src/image_model/downsampling_module.h
@@ -15,7 +15,7 @@ class DownsamplingModule : public DegradationOperator {
  public:
   // The given scale parameter is the scale by which the resized image will
   // be modified. The scale should be greater than or equal to 1.
-  explicit DownsamplingModule(const double scale);
+  explicit DownsamplingModule(const int scale);
 
   virtual void ApplyToImage(ImageData* image_data, const int index) const;
 
@@ -23,7 +23,7 @@ class DownsamplingModule : public DegradationOperator {
       const cv::Size& image_size, const int index) const;
 
  private:
-  const double scale_;
+  const int scale_;
 };
 
 }  // namespace super_resolution
